+++ src/domain/fetchJson.ts
@@ -1,3 +1,26 @@
+export type FetchedMovieItemJson = {
+  adult: boolean;
+  backdrop_path: string;
+  genre_ids: number[];
+  id: number;
+  original_language: string;
+  original_title: string;
+  overview: string;
+  popularity: number;
+  poster_path: string;
+  release_date: string;
+  title: string;
+  video: boolean;
+  vote_average: number;
+  vote_count: number;
+};
+
+export type FetchedMovieJson = {
+  page: number;
+  results: FetchedMovieItemJson[];
+  total_pages: number;
+};
+
 type FetchJson = <T>(url: string) => Promise<T>;
 
 const fetchJson: FetchJson = async api => {
