+++ src/domain/processMovieData.ts
@@ -1,8 +1,14 @@
-import { MovieFetchedJson, Movie } from '../app';
+import { FetchedMovieJson } from './fetchJson';
+
+export type Movie = {
+  title: string;
+  backdropPath: string;
+  voteAverage: number;
+};
 
 const MOVIE_IMAGE_URL = 'https://image.tmdb.org/t/p/w220_and_h330_face/';
 
-const processMovieData = ({ page, results, total_pages: totalPages }: MovieFetchedJson) => {
+const processMovieData = ({ page, results, total_pages: totalPages }: FetchedMovieJson) => {
   const movies: Movie[] = results.map(result => ({
     title: result.title,
     backdropPath: `${MOVIE_IMAGE_URL}${result.backdrop_path}`,
